cleanup: move to separate file, config updates
